backend/requirements.txt
# Upstream chapter13-aligned backend dependencies

# HelloAgents framework and protocol support
hello-agents[protocols]>=0.2.4,<=0.2.9
fastmcp>=2.0.0
uv>=0.8.0

# FastAPI and app runtime
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
loguru>=0.7.0

# HTTP clients used by the current local code
requests>=2.31.0
httpx>=0.27.0
aiohttp>=3.10.0

# LLM client used by the current local code
openai>=1.0.0

# Utilities
python-dateutil>=2.8.2
huggingface_hub>=0.25.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
